refactor: update imports and remove unused registries for improved clarity and maintainability in type handling
